use i64vec3 instead of ivec3

+++ include/generation/UniversalPosition.hpp
@@ -12,9 +12,9 @@ private:
 public:
     uint32_t seedUniverse;
     glm::ivec3 positionGalaxy;
-    glm::ivec3 positionSolarSystem;
+    glm::i64vec3 positionSolarSystem;
 
-    UniversalPosition(uint32_t seedUniverse, glm::ivec3 galaxy, glm::ivec3 solarSystem);
+    UniversalPosition(uint32_t seedUniverse, glm::ivec3 galaxy, glm::i64vec3 solarSystem);
     UniversalPosition(uint32_t seedUniverse, glm::ivec3 galaxy);
     UniversalPosition(uint32_t seedUniverse);
     UniversalPosition();
